No need to call setup, add toggle-mirrored.
The kinect object is now initialized when it is first needed, so there
is no need to call `bifocals.core/setup` before using other functions.
However, you can continue to call it, in case you don't use it right
away and want to avoid a long delay in the middle of your sketch, while
the kinect initializes.

Add a function in core to toggle whether the kinect's depth and scene
images are mirrored.